+++ cli/src/commands/init.rs
@@ -12,15 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::io;
 use std::io::Write;
-use std::path::Path;
 
 use clap::ArgGroup;
-use itertools::Itertools as _;
 use jj_lib::file_util;
-use jj_lib::repo::Repo;
-use jj_lib::view::View;
 use jj_lib::workspace::Workspace;
 use tracing::instrument;
 
@@ -59,7 +54,7 @@ pub(crate) fn cmd_init(
         .map_err(|e| user_error_with_message("Failed to create workspace", e))?;
 
     if args.git || args.git_repo.is_some() {
-        git_init(ui, command, &wc_path, args.git_repo.as_deref())?;
+        git::git_init(ui, command, &wc_path, args.git_repo.as_deref())?;
     } else {
         if !command.settings().allow_native_backend() {
             return Err(user_error_with_hint(
@@ -77,88 +72,5 @@ Set `ui.allow-init-native` to allow initializing a repo with the native backend.
         "Initialized repo in \"{}\"",
         relative_wc_path.display()
     )?;
-    Ok(())
-}
-
-fn print_trackable_remote_branches(ui: &Ui, view: &View) -> io::Result<()> {
-    let remote_branch_names = view
-        .branches()
-        .filter(|(_, branch_target)| branch_target.local_target.is_present())
-        .flat_map(|(name, branch_target)| {
-            branch_target
-                .remote_refs
-                .into_iter()
-                .filter(|&(_, remote_ref)| !remote_ref.is_tracking())
-                .map(move |(remote, _)| format!("{name}@{remote}"))
-        })
-        .collect_vec();
-    if remote_branch_names.is_empty() {
-        return Ok(());
-    }
-
-    writeln!(
-        ui.hint(),
-        "The following remote branches aren't associated with the existing local branches:"
-    )?;
-    let mut formatter = ui.stderr_formatter();
-    for full_name in &remote_branch_names {
-        write!(formatter, "  ")?;
-        writeln!(formatter.labeled("branch"), "{full_name}")?;
-    }
-    drop(formatter);
-    writeln!(
-        ui.hint(),
-        "Hint: Run `jj branch track {names}` to keep local branches updated on future pulls.",
-        names = remote_branch_names.join(" "),
-    )?;
-    Ok(())
-}
-
-// TODO(essiene): Move to cli/src/commands/git.rs for `jj git init`
-fn git_init(
-    ui: &mut Ui,
-    command: &CommandHelper,
-    workspace_root: &Path,
-    git_repo: Option<&str>,
-) -> Result<(), CommandError> {
-    let cwd = command.cwd().canonicalize().unwrap();
-    let relative_wc_path = file_util::relative_path(&cwd, workspace_root);
-
-    if let Some(git_store_str) = git_repo {
-        let git_store_path = cwd.join(git_store_str);
-        let (workspace, repo) =
-            Workspace::init_external_git(command.settings(), workspace_root, &git_store_path)?;
-        let mut workspace_command = command.for_loaded_repo(ui, workspace, repo)?;
-        git::maybe_add_gitignore(&workspace_command)?;
-        // Import refs first so all the reachable commits are indexed in
-        // chronological order.
-        workspace_command.import_git_refs(ui)?;
-        workspace_command.maybe_snapshot(ui)?;
-        if !workspace_command.working_copy_shared_with_git() {
-            let mut tx = workspace_command.start_transaction();
-            jj_lib::git::import_head(tx.mut_repo())?;
-            if let Some(git_head_id) = tx.mut_repo().view().git_head().as_normal().cloned() {
-                let git_head_commit = tx.mut_repo().store().get_commit(&git_head_id)?;
-                tx.check_out(&git_head_commit)?;
-            }
-            if tx.mut_repo().has_changes() {
-                tx.finish(ui, "import git head")?;
-            }
-        }
-        print_trackable_remote_branches(ui, workspace_command.repo().view())?;
-    } else {
-        if workspace_root.join(".git").exists() {
-            return Err(user_error_with_hint(
-                "Did not create a jj repo because there is an existing Git repo in this directory.",
-                format!(
-                    r#"To create a repo backed by the existing Git repo, run `jj init --git-repo={}` instead."#,
-                    relative_wc_path.display()
-                ),
-            ));
-        }
-
-        Workspace::init_internal_git(command.settings(), workspace_root)?;
-    }
-
     Ok(())
 }
